2007-11-23 18.00 UTC+0700 Andi Jahja <harbour/AT/cbn/net/id>    * source/commond.c      ! commented unnecessary hb_retl().
   * source/c_ipaddr.c
   * source/richedit.c
   * source/procmisc/blowfish.c
     ! castings

   * source/resource.c
     ! modified pragma startup to init procedure.
       (actually a work-around for DMC, but it seems to work for others)

   * source/window.c
     ! commented unused variables


git-svn-id: https://svn.code.sf.net/p/hwgui/code/trunk/hwgui@815 66345640-f411-406e-acba-706ed0d45c9c

+++ source/c_ipaddr.c
@@ -107,10 +107,10 @@ HB_FUNC( GETIPADDRESS )
 
 	SendMessage(hWnd, IPM_GETADDRESS, 0, (LPARAM)(LPDWORD)&pdwAddr);
 
-	v1 = FIRST_IPADDRESS( pdwAddr );
-	v2 = SECOND_IPADDRESS( pdwAddr );
-	v3 = THIRD_IPADDRESS( pdwAddr );
-	v4 = FOURTH_IPADDRESS( pdwAddr );
+	v1 = (BYTE) FIRST_IPADDRESS( pdwAddr );
+	v2 = (BYTE) SECOND_IPADDRESS( pdwAddr );
+	v3 = (BYTE) THIRD_IPADDRESS( pdwAddr );
+	v4 = (BYTE) FOURTH_IPADDRESS( pdwAddr );
 
 	hb_reta( 4 );
 	hb_storni( (INT) v1, -1, 1 );
